Refactor: sandbox logic into pkg/sandbox
Moves the sandbox management and execution logic from pkg/commands into a dedicated pkg/sandbox package to improve reusability and separation of concerns.

- Add local and pod executor interface
- Create pkg/sandbox package with interactive.go, exec.go, and types.go.
- Move CodebotSandbox (renamed to InteractiveSandbox) and related methods.
- Move pod execution helpers (ExecInPod, WriteFileInPod, etc.) and thread data structures.
- Update github-fix-issue, github-feedback, github-autopoll, threads, code, and tmux commands to consume the new package.
- Remove pkg/commands/sandbox.go.
- Fix linting issues: rename SandboxNameForIssue to NameForIssue to avoid stuttering, and remove unnecessary type conversions.
