+++ repo-agent/pkg/commands/github-fix-issue.go
@@ -1,23 +1,16 @@
 package commands
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"io"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/pkg/clients"
 	"github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/pkg/github"
 	"github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/pkg/prompts"
+	"github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/pkg/sandbox"
 	"github.com/spf13/cobra"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/tools/remotecommand"
 	"k8s.io/klog/v2"
 )
 
@@ -81,200 +74,77 @@ func RunGithubFixIssue(ctx context.Context, opt GithubFixIssueOptions) error {
 		return fmt.Errorf("failed to generate prompt for issue: %w", err)
 	}
 
-	sandbox, found, err := findSandboxForIssue(ctx, kube, repo, issue)
+	sb, found, err := sandbox.FindSandboxForIssue(ctx, kube, repo, issue)
 	if err != nil {
 		return err
 	}
 
 	if !found {
-		sandbox, err = launchSandboxForIssue(ctx, kube, repo, issue)
+		sb, err = sandbox.LaunchSandboxForIssue(ctx, kube, repo, issue)
 		if err != nil {
 			return fmt.Errorf("launching sandbox for issue: %w", err)
 		}
 	}
 
-	geminiAPIKey, err := GetGeminiAPIKey(sandbox.podID.Namespace + "/" + sandbox.podID.Name)
+	podID := sb.GetPodID()
+
+	geminiAPIKey, err := GetGeminiAPIKey(podID.Namespace + "/" + podID.Name)
 	if err != nil {
 		return err
 	}
 
-	if err := sandbox.setupGit(ctx); err != nil {
+	if err := sb.SetupGit(ctx); err != nil {
 		return fmt.Errorf("setting up git in sandbox: %w", err)
 	}
 
-	if err := sandbox.SetupGitRepos(ctx); err != nil {
+	if err := sb.SetupGitRepos(ctx); err != nil {
 		return fmt.Errorf("setting up git branches in sandbox: %w", err)
 	}
 
 	// Copy the prompt into the pod (for now)
 	if len(prompt) > 0 {
-		log.Info("copying prompt into sandbox pod", "pod", sandbox.podID)
+		log.Info("copying prompt into sandbox pod", "pod", podID)
 
 		path := "/workspaces/prompt.txt"
-		if err := writeFileInPod(ctx, kube, sandbox.podID, path, prompt); err != nil {
+		if err := sb.WriteFile(ctx, path, prompt); err != nil {
 			return fmt.Errorf("copying prompt into sandbox pod: %w", err)
 		}
 
-		log.Info("Copied prompt into sandbox pod", "pod", sandbox.podID, "path", path)
+		log.Info("Copied prompt into sandbox pod", "pod", podID, "path", path)
 	}
 
 	// HACK: Avoid git lock issues
 	time.Sleep(5 * time.Second)
 
-	if err := sandbox.CheckoutNewBranch(ctx); err != nil {
+	if err := sb.CheckoutNewBranch(ctx); err != nil {
 		return fmt.Errorf("checking out branch: %w", err)
 	}
 
-	if err := configureGemini(ctx, sandbox); err != nil {
+	if err := sandbox.ConfigureGemini(ctx, sb); err != nil {
 		return fmt.Errorf("configuring gemini in sandbox: %w", err)
 	}
 
 	// Run gemini with API key and prompt
 	{
-		log.Info("Running gemini in pod", "pod", sandbox.podID)
+		log.Info("Running gemini in pod", "pod", podID)
 
-		workdir := fmt.Sprintf("/workspaces/%s", sandbox.repo.FilesystemName())
+		workdir := fmt.Sprintf("/workspaces/%s", sb.GetRepo().FilesystemName())
 
 		// TODO:
 		// export GEMINI_TELEMETRY_ENABLED=true
 		// export GEMINI_TELEMETRY_OTLP_ENDPOINT=http://otel-portal.otel-system:4317
 
-		opts := execOptions{
+		opts := sandbox.ExecOptions{
 			Command: []string{"sh", "-c", fmt.Sprintf("cd %s && export GEMINI_API_KEY=%s && gemini --yolo --model gemini-3-pro-preview < /workspaces/prompt.txt", workdir, geminiAPIKey)},
 			Stdout:  os.Stdout,
 			Stderr:  os.Stderr,
 		}
 		opts.Secrets = []string{geminiAPIKey}
 
-		if err := execInPod(ctx, kube, sandbox.podID, opts); err != nil {
+		if err := sb.Exec(ctx, opts); err != nil {
 			return fmt.Errorf("running gemini: %w", err)
 		}
 	}
 
 	return nil
-}
-
-type execOptions struct {
-	Command []string
-	Secrets []string
-
-	Stdin  []byte
-	Stdout io.Writer
-	Stderr io.Writer
-}
-
-// execInPod writes the specified data to a file in the specified pod.
-func execInPod(ctx context.Context, kube *clients.KubernetesClient, podID types.NamespacedName, opts execOptions) error {
-	log := klog.FromContext(ctx)
-
-	redactedCommand := strings.Join(opts.Command, " ")
-	for _, v := range opts.Secrets {
-		redactedCommand = strings.ReplaceAll(redactedCommand, v, "****")
-	}
-
-	podExecOptions := &v1.PodExecOptions{
-		// Container: containerName,
-		Command: opts.Command,
-		Stdin:   true,
-		Stdout:  true,
-		Stderr:  true,
-		TTY:     false,
-	}
-	if opts.Stdin == nil {
-		podExecOptions.Stdin = false
-	}
-
-	req := kube.Clientset.CoreV1().RESTClient().Post().
-		Resource("pods").
-		Name(podID.Name).
-		Namespace(podID.Namespace).
-		SubResource("exec").
-		VersionedParams(podExecOptions, scheme.ParameterCodec)
-
-	url := req.URL().String()
-	exec, err := remotecommand.NewWebSocketExecutor(kube.RestConfig, "POST", url)
-	if err != nil {
-		return fmt.Errorf("executing command in pod: %w", err)
-	}
-
-	var stdout bytes.Buffer
-	var stderr bytes.Buffer
-
-	streamOptions := remotecommand.StreamOptions{
-		Stdout: &stdout,
-		Stderr: &stderr,
-		Tty:    false,
-	}
-	if opts.Stdin != nil {
-		streamOptions.Stdin = bytes.NewReader(opts.Stdin)
-	}
-	if opts.Stdout != nil {
-		streamOptions.Stdout = opts.Stdout
-	}
-	if opts.Stderr != nil {
-		streamOptions.Stderr = opts.Stderr
-	}
-
-	// Run the command
-	if err := exec.StreamWithContext(ctx, streamOptions); err != nil {
-		log.Error(err, "executing command", "pod", podID, "command", redactedCommand, "stdout", stdout.String(), "stderr", stderr.String())
-		return fmt.Errorf("streaming command in pod: %w", err)
-	}
-
-	log.Info("executed command", "pod", podID, "command", redactedCommand, "stdout", stdout.String(), "stderr", stderr.String())
-	return nil
-}
-
-// writeFileInPod writes the specified data to a file in the specified pod.
-func writeFileInPod(ctx context.Context, kube *clients.KubernetesClient, podID types.NamespacedName, path string, data []byte) error {
-	// log := klog.FromContext(ctx)
-
-	var stdout bytes.Buffer
-
-	opt := execOptions{
-		Command: []string{"/bin/tee", path},
-		Stdin:   data,
-		Stdout:  &stdout, // To avoid logging to stdout
-	}
-
-	return execInPod(ctx, kube, podID, opt)
-}
-
-// waitForPodReady waits for the specified pod to be ready.
-func waitForPodReady(ctx context.Context, kube *clients.KubernetesClient, podID types.NamespacedName) error {
-	log := klog.FromContext(ctx)
-
-	clientset := kube.Clientset
-
-	log.Info("Waiting for sandbox pod to be ready", "pod", podID.Name)
-
-	ctx, cancel := context.WithTimeout(ctx, 2*time.Minute)
-	defer cancel()
-
-	for {
-		stream, err := clientset.CoreV1().Pods(podID.Namespace).Watch(ctx, metav1.ListOptions{FieldSelector: "metadata.name=" + podID.Name, Watch: true})
-		if err != nil {
-			return err
-		}
-		defer stream.Stop()
-		for event := range stream.ResultChan() {
-			pod, ok := event.Object.(*v1.Pod)
-			if !ok {
-				return fmt.Errorf("unexpected type %T when watching pod", event.Object)
-			}
-			if isPodReady(pod) {
-				log.Info("Sandbox pod is ready", "pod", podID.Name)
-				return nil
-			}
-		}
-	}
-}
-
-func isPodReady(pod *v1.Pod) bool {
-	for _, cond := range pod.Status.Conditions {
-		if cond.Type == v1.PodReady {
-			return cond.Status == v1.ConditionTrue
-		}
-	}
-	return false
 }
