+++ repo-agent/pkg/commands/helpers.go
@@ -7,53 +7,10 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/pkg/clients"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/pkg/sandbox"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/klog/v2"
 )
-
-// findSandboxPod finds the pod for the given sandbox name.
-// If the pod is not found, it returns (nil, nil)
-func findSandboxPod(ctx context.Context, sandboxName string) (*types.NamespacedName, error) {
-	kube, err := clients.NewKubernetesClient()
-	if err != nil {
-		return nil, err
-	}
-
-	clientset := kube.Clientset
-	namespace := kube.CurrentNamespace
-
-	// The sandbox name in the RGD is devc-<name>
-	// And the pods have label sandbox=devc-<name>
-	labelSelector := fmt.Sprintf("sandbox=devc-%s", sandboxName)
-	pods, err := clientset.CoreV1().Pods(namespace).List(ctx, metav1.ListOptions{
-		LabelSelector: labelSelector,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to list pods with selector %q: %w", labelSelector, err)
-	}
-
-	if len(pods.Items) == 0 {
-		return nil, nil
-	}
-
-	// Pick the first running pod, or just the first one if none are running yet (though exec will fail)
-	pod := &pods.Items[0]
-
-	for _, p := range pods.Items {
-		if p.Status.Phase == "Running" {
-			pod = &p
-			break
-		}
-	}
-
-	podID := &types.NamespacedName{
-		Name:      pod.Name,
-		Namespace: pod.Namespace,
-	}
-	return podID, nil
-}
 
 // updateSSHConfig updates the user's SSH config to include the sandbox host.
 func updateSSHConfig(ctx context.Context, path, host string, podID types.NamespacedName) error {
@@ -75,7 +32,7 @@ func updateSSHConfig(ctx context.Context, path, host string, podID types.Namespa
 	block := []string{
 		"# Generated by gemini-sandbox",
 		fmt.Sprintf("Host %s", host),
-		fmt.Sprintf("  ProxyCommand kubectl exec -i %s -n %s -- %s sshd", podID.Name, podID.Namespace, repoSandboxBinary),
+		fmt.Sprintf("  ProxyCommand kubectl exec -i %s -n %s -- %s sshd", podID.Name, podID.Namespace, sandbox.RepoSandboxBinary),
 		"  StrictHostKeyChecking no",
 		"  UserKnownHostsFile /dev/null",
 	}
